Avoid name duplication by renaming main modules to "index"
