+++ src/FollowReveal/docs/index.mdx
@@ -13,7 +13,7 @@ import { Props, Playground } from "docz";
 import { Front, PlaceHolderImage } from "docs/PlayGroundUtils";
 import Grid from "./Grid";
 
-import FollowReveal from "../FollowReveal";
+import FollowReveal from "../";
 
 <Props of={FollowReveal} />
