+++ src/TurnReveal/docs/index.mdx
@@ -20,7 +20,7 @@ import {
 } from "docs/PlayGroundUtils";
 import ControlWrapper from "./ControlWrapper";
 
-import TurnReveal from "../TurnReveal";
+import TurnReveal from "../";
 
 <Props of={TurnReveal} />
