+++ src/index.js
@@ -1,4 +1,4 @@
-export default from "./TurnReveal/TurnReveal";
-export FollowReveal from "./FollowReveal/FollowReveal";
+export default from "./TurnReveal";
+export FollowReveal from "./FollowReveal";
 export Direction from "./Direction";
 export Transition from "./Transition";
